@@ -31,10 +31,10 @@
         }
     },
     "require": {
-        "php": "^7.1 || ^8.0",
+        "php": "^7.2 || ^8.0",
         "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
-        "slevomat/coding-standard": "^7.0.1",
-        "squizlabs/php_codesniffer": "^3.6.0"
+        "slevomat/coding-standard": "^8",
+        "squizlabs/php_codesniffer": "^3.7"
     },
     "require-dev": {
         "symfony/phpunit-bridge": "^5.1"
